Removed allot of casting from the code. Both using generics, and correct typing.

+++ projects/client/RabbitMQ.Client/src/client/api/IContentHeader.cs
@@ -50,7 +50,7 @@ namespace RabbitMQ.Client
         /// <summary>
         /// Retrieve the AMQP class ID of this content header.
         /// </summary>
-        int ProtocolClassId { get; }
+        ushort ProtocolClassId { get; }
 
         /// <summary>
         /// Retrieve the AMQP class name of this content header.
